news_event/requirements.txt
feedparser>=6.0
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.0
pyyaml>=6.0
anthropic>=0.89
